Lighting Maps
Basically, the ambient, diffuse and specular components can't be
the same for the object across it as it can be made up of different
materials. Thus we use diffuse map, specular maps to define different
lighting behaviour. We can also have emission maps which can simulate
glow from the object itself.

Signed-off-by: Sai Teja <saiteja13427@gmail.com>