Blending
Blending is basically used for rendering transparent objects.

The easiest kind of blending is where some fragments are totally
transparent in a texture and they can be discarded in fragment shader
and the other fragments can be kept.

Then when we have semi transparent textures, we have to enable blending
in  OpenGL.

glEnable(GL_BLEND);

Then OpenGL provides some functions to select how to compute the resulting color
in the transparent places of the texture.

It generally uses the following formula

color = source color * source factor + destination color * destination factor.

We have to set source factor and destination factor using the function

glBlendFunc(source factor, dst factor);

The most used arguments being GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA, which
means we multiple source (texture) color with its alpha value and add it to
destination color (existing color in color buffer) times (1 - source alpha).

Note: When blending transparent objects keeping the objects ordered from farthest
to nearest is necessary else we see weird depth test related issues because depth test
doesn't check if a fragment is transparent or not, if there is a transparent fragment in
the nearest texture, all the fragments behind it in farther objects will not be rendered
which will lead to a weird effect.

So, when we have transparent objects,
1. First render all opaque objects.
2. Sort the transparent objects from farthest to closest. (This gets difficult if we have random objects.)
3. Render the transparent objects in the same order.

Signed-off-by: Sai Teja <saiteja13427@gmail.com>