Framebuffers
We have a by default framebuffer usually created by GLFW.
We can create a new framebuffer ourselved attach a texture
to it for color buffer and a render buffer for depth and stencil
for it.

Then we can bind to it and draw to that framebuffer. We can
draw the whole scene to a single texture using this new frame-
buffer and we can then do interesting postprocessing effects
using different kernels.

Kernel are 3x3 matrices which we use to multiply the surrounding
fragments and add the colors to current fragment helping us
create postprocessing effects.

Signed-off-by: Sai Teja <saiteja13427@gmail.com>