Cubemaps
Cubemaps is an interesting concept wherein there are
multiple textures mapped to one. For example there
is a texture with 6 faces which can form a cube of textures
called the cubemap.

There are pretty interesting applications for this, one such thing
is the skybox. Skybox is when you are placed within a cube with a cubemap
around giving a feeling that you are inside a large scene.

With this skybox, we can create interesting reflecting, refracting etc
behvaiours on the objects inside by sampling the reflected, refracted fragments
and putting them onto the objects inside.

Signed-off-by: Sai Teja <saiteja13427@gmail.com>