Advanced GLSL
GLSL has some more advanced concepts like the below

1. Vertex shaders have some interesting builtin variables like
gl_VertexID, gl_PointSize.

2. Fragment shader have some useful building variables like
gl_FragCoord, gl_FrontFacing, gl_FragDepth.

3. GLSL has interface blocks where we can group a lot of
in/out variables to pass between stages of shaders.

4. Uniform buffer objects is a way to a few uniforms
globally which can be accessible by any shader irrespective
of where it was in use or not at the time of setting. Though this
involves a bit of heavylifting wherein we will have to query the bindpoint
of each buffer object in each shader and then bind it to required
index. We will also have to create uniform buffers, calculate the
offsets of each member and fill it accordingly using glBufferData
or glBufferSubData.

Signed-off-by: Sai Teja <saiteja13427@gmail.com>