Geometry Shader
Geometry shader is a wellknow shader stage. It is basically used
to take the primitives from vertex shader and can be used to output
additional primitives or totally different primitives.

This has applications like
1. It can be used to make a exploding effect by traslating the primitives
on its normal.

2. We can visualize the normals of each vertex which can be used for
debugging purpose which can help us to see if the supplied vertex data
is correct or not.

Signed-off-by: Sai Teja <saiteja13427@gmail.com>