Instancing
When we start drawing 1000s and 10000s of objects in our scenes, we
can quickly run into problems related to slow communication speeds
between CPU and GPU.

GPU is super fast with rendering, but if we do render calls 1000s of times,
each time CPU have to send in data to GPU and GPU have to draw it. This
is a major bottlenech.

Thus, we have instanced draw methods like glDrawElementsInstanced and
glDrawElementsInstanced. These instanced calls take a extra param as to
how many times to draw the primitive.

Now drawing 100 times normally will lead to the same objects being rendered
at same place for 100 times which is of no use. We thus have to have
a position offset array which we transfer to vertexShader. The vertex shader
would then pick the respective offset using the builtin gl_InstanceID variable
which tells which instance are we currently rendering.

We can either pass this offset array using uniforms or instanced arrays. Uniforms
have a limit on how much uniform data we can send, thus, we have instanced arrays
which are essentially array buffers linked to some layout location in vertex shader.
We just have to tell opengl that these are instanced arrays and we need to change
to the next element when an instance is done by using glVertexAttribDivisor and set
that layouts divisor to 1 which would tell that its an instance array and not normal
input.

Signed-off-by: Sai Teja <saiteja13427@gmail.com>