Anti Aliasing
While rendering edges, we generally see jagged edges many times.
Jagged edges decreases the quality of the rendered scene drastically.

To solve this we have multiple anti aliasing techniques.

1. Initially we had an anti-aliasing technique called super sample anti-
aliasing. In here we do the rendering in a higher resolution and then while drawing
the full scene we decrease the resolution which would eliminate the jagged edges.
This is not very good efficiency wise but it gave birth to a more efficient
approach called MSAA.

2. MSAA (Multisample anti aliasing) is a technique where we use multiple points in a
pixes to tell if the boundary overlaps or not. The more the points inside the boundary
the darker the color. With this approach we have lighter colors on the edges bleding with
the actual object. This covers up the jagged up edges pretty well.
This is mostly taken care by the rasterizer in the driver, we just need to have a multisample
framebuffer and enable this feature in opengl.

GLFW window with multi sampled framebuffer: glfwWindowHint(GLFW_SAMPLES, 4);
Enable multisample in opengl: glEnable(GL_MULTISAMPLE);

Signed-off-by: Sai Teja <saiteja13427@gmail.com>